changement dans deg_to_rad

+++ minirt.h
@@ -226,7 +226,7 @@ t_float3 *ft_calculate_viewport(t_list_minirt *camera);
 t_float3 ft_canvas_to_viewport(int x, int y, t_list_minirt *camera, t_list_minirt *resolution);
 float ft_closest_intersection(t_float3 origin, t_float3 direction, float tmin, float tmax, t_list_minirt *head, t_list_minirt **closest_sphere);
 float ft_compute_lighting(t_float3 p, t_float3 n, t_list_minirt *head,  float tmin, float tmax);
-int	ft_deg_to_rad(int deg);
+float	ft_deg_to_rad(int deg);
 float    ft_dot_product(t_float3 a, t_float3 b);
 t_list_minirt *ft_find_next_ambient(t_list_minirt *previous_light, t_list_minirt *head);
 t_list_minirt *ft_find_next_light(t_list_minirt *previous_light, t_list_minirt *head);
