+++ raytracing/ft_deg_to_rad.c
@@ -1,6 +1,6 @@
 #include "../minirt.h"
 
-int	ft_deg_to_rad(int deg)
+float	ft_deg_to_rad(int deg)
 {
 	return (deg * 2 * M_PI / 360);
 }
